Separate IWalletFeePolicy into its own file
